+++ packages/ui/core-components/src/lib/atoms/inputs/dropdown/Dropdown.svelte
@@ -202,6 +202,8 @@
 				if ($option.removeOnDeselect) flagOption([$option, DropdownValueFlag.REMOVE_ON_DESELECT]);
 			});
 			queryOptions = query;
+			hasHadSelection = false;
+			optionUpdates = undefined;
 		}
 	}, 250);
 
@@ -222,7 +224,6 @@
 			if (!hasHadSelection) {
 				setTimeout(evalDefaults, 0);
 				optionUpdates();
-				optionUpdates = undefined;
 			}
 		});
 	}
@@ -291,6 +292,11 @@
 			flagOption([opt, DropdownValueFlag.FORCE_SELECT]);
 		});
 	}
+
+	function getIdx(queryOpt) {
+		if ('similarity' in queryOpt) return queryOpt.similarity * -1;
+		return queryOpt.ordinal ?? 0;
+	}
 </script>
 
 <slot />
@@ -300,7 +306,7 @@
 		<DropdownOption
 			value={queryOpt.value}
 			valueLabel={queryOpt.label}
-			idx={(queryOpt.similarity ?? 0) * -1 ?? -1}
+			idx={getIdx(queryOpt)}
 			__auto
 		/>
 	{/each}