+++ packages/ui/core-components/src/lib/unsorted/viz/map/EvidenceMap.js
@@ -13,10 +13,16 @@ export class EvidenceMap {
 	/** @type {import('leaflet').Map | undefined} */
 	#map;
 
-	/** Manages the last clicked marker to track selection states. */
+	/**
+	 * Manages the last clicked marker to track selection states.
+	 * @type {import("leaflet").Marker}
+	 */
 	#lastClickedMarker = null;
 
-	/** Stores original styles of markers to restore after changes. */
+	/**
+	 * Stores original styles of markers to restore after changes.
+	 * @type {Map<import("leaflet").Marker>}
+	 */
 	#markerStyles = new Map();
 
 	/** @type {HTMLDivElement | undefined} */
@@ -25,7 +31,10 @@ export class EvidenceMap {
 	/** Handles the promises associated with the initialization of the map component. */
 	#sharedPromise = sharedPromise();
 
-	/** Manages the bounds of the map to adjust the view based on the layers added. */
+	/**
+	 * Manages the bounds of the map to adjust the view based on the layers added.
+	 * @type {import('leaflet').Bounds}
+	 */
 	#bounds;
 
 	/** Tracks whether the initial view has been set based on data bounds. */
@@ -53,11 +62,17 @@ export class EvidenceMap {
 	}
 
 	/**
+	 * @type {number}
+	 */
+	#initZoom;
+
+	/**
 	 * Initializes the map within the provided HTML element with specified starting coordinates and zoom level.
 	 * @param {HTMLDivElement} mapEl - The HTML element to initialize the map in.
 	 * @param {string} basemap - The URL template for the basemap.
-	 * @param {import('leaflet').LatLngExpression} startingCoords - The starting coordinates for the map.
-	 * @param {number} startingZoom - The starting zoom level for the map.
+	 * @param {import('leaflet').LatLngExpression} [startingCoords] - The starting coordinates for the map.
+	 * @param {number} [startingZoom] - The starting zoom level for the map.
+	 * @param {boolean} [userDefinedView=false]
 	 * @returns {Promise<void>}
 	 */
 	async init(mapEl, basemap, startingCoords, startingZoom, userDefinedView = false) {
@@ -70,6 +85,8 @@ export class EvidenceMap {
 				});
 		}
 
+		this.#initZoom = startingZoom;
+
 		if (this.#mapEl) {
 			const e = new Error('Evidence Map already initialized');
 			this.#sharedPromise.reject(e);
@@ -77,9 +94,9 @@ export class EvidenceMap {
 		}
 
 		this.#mapEl = mapEl;
-		this.#map = Leaflet.map(this.#mapEl, { zoomControl: false }).setView(
+		this.#map = Leaflet.map(this.#mapEl, { zoomControl: false, zoomSnap: 0.25 }).setView(
 			startingCoords,
-			startingZoom
+			startingZoom ?? 1
 		);
 		if (userDefinedView) {
 			this.#initialViewSet = true; // Mark initial view as set
@@ -121,7 +138,8 @@ export class EvidenceMap {
 		});
 
 		if (this.#bounds.isValid()) {
-			this.#map.fitBounds(this.#bounds);
+			this.#map.fitBounds(this.#bounds, { maxZoom: 12 });
+			if (this.#initZoom) this.#map.setZoom(this.#initZoom);
 		} else {
 			console.error('Bounds are invalid!', this.#bounds);
 			throw new Error('Bounds are invalid!');
@@ -191,6 +209,8 @@ export class EvidenceMap {
 		this.#bounds.extend(geoJsonLayer.getBounds());
 		if (!this.#initialViewSet) {
 			this.updateBounds();
+		} else {
+			this.#map.setZoom(this.#initZoom ?? 5);
 		}
 		geoJsonLayer.bringToBack();
 		return geoJsonLayer;
@@ -254,6 +274,8 @@ export class EvidenceMap {
 		if (coords && Array.isArray(coords) && coords.length === 2 && !this.#initialViewSet) {
 			this.#bounds.extend(coords);
 			this.updateBounds();
+		} else {
+			this.#map.setZoom(this.#initZoom ?? 5);
 		}
 
 		return marker;